skeleton for restful webservice using scala2.9/akka 1.1.2/jetty 8.
presently builds and runs according to the README.  

project-specific code is present on paper, needs to be integrated here.

the checkin directory is a bit verbose, and due to some mis-clicks in eclipse, i have a bunch of class files checked in as well.  will clean this up later, but i don't want to get bogged down in the details just yet.