Can now call RegistryClient to get repository EPR.  Agent responds to this as: /agent/<UUID>/repository-epr
Spent some time hacking the CachedEPR class, which is still buggy. Not a priority item for the demo tomorrow, will be revisited.

Upped scala version in build to 2.9.0-1 while trying to figure out why some interim bad code was bad.  I fixed the bad code, will keep the higher version, since it works.

Added a script to manually remove the conflicting slf4j-api-1.6.0 jar which seems to be auto-grabbed when I do sbt update.  See if I can stop that grab in the build props.

Cassandra and Solr paths in registry are currently hardcoded, read that from config.properties.