Description of previous commit: Added the polling method to agent, and the REST hook for it.
Lots of cleanup...

curl_tests/simple_tests.rb now polls the algorithm status after 0,5, and 30 seconds.  We do manage to artificially reach a 'finished' status at this point, and the test picks that up.

It takes quite a while for the volume IDs to come out of the registry... and, of course, there are a lot of them.

We'll worry about that timing delay later.  If I'm lucky, I can actually call an executable and get it to run...