cleaned up functions for getting the algorithm out of registry.  this is far from perfect, but we presently copy the algorithm executable to the working directory.
removed java setProperty("user.dir") calls from the code.  we now attempt use a launch script (launch-app.sh) as a subprocess to launch jars downloaded from the registry.  this script takes the working directory as a parameter so that we can avoid changing working dir within the Agent code.

having read the java Runtime.exec documentation more thoroughly, i see that the above step is not really necessary: we can pass the working directory as a parameter to Runtime.exec

added call to pass "volumelist=SOMENAME.txt" as a parameter in the properties file used by the algorithm executable

we now write the volume list out to a text file in the working directory.