Workaround for registry flakiness documented in: http://wso2.org/forum/thread/14971
We can now set a flag in config.properties that allows the agent to use a locally-cached copy of an executable, instead of trying to fetch it from the registry.  This is a stopgap measure that should only be needed until we have the above problem solved.

To enable this option, set the following in config.properties:

skipgettingalgoexecutablefromregistry=true

You'll need the correct algo executable file to be placed in the "emergency-algorithm-executable" directory of the project root.