couple of quick unit tests for the replaceEscaped function.  writing out the properties file is currently problematic and is complicated by the fact that java's escaping of things like backslash is compounded by java regexp library escaping of things like backslash.  so to match a single backslash in a regexp, you need to write "\\\\".
i tested this thoroughly against Jiaan's jar file and the properties file being bumped with entries like:

http\://coffeetree.cs.indiana.edu\:9992

causes that jar to crash.  this isn't a problem with the jar, it's a problem with how the agent dumps the necessary properties file.  running out of thinking-power for the day, but i'll need to dig into this and get it solved prior to the deadline.

having unit tests up and running with SBT makes this a much easier problem to solve.

-f