Added methods to FirstRegistry for posting results to registry.
Seems to work, but a bit rough.  Here are the main concerns:

  Registry API doesn't let us use generally accepted URIs such as 
  
  /results/urn:publicid:Felix+A1+Happy/my-algo-run-6123/stdout.txt
  /results/urn:publicid:Felix%2dEscaped%4dURI/my-algo-run-555/some-file-result.txt
  
So, as a workaround, I convert the urn identifier for a user's name to a hex string.  This uses org.apache.commons.codec.binary.Hex, and while it's not easy for a human to read, it gives us a 1:1 mapping that we don't have to explicitly track.  I'm open to a conversation about better ways to do this, but I need to push on, so for now this is a compromise.

Thus, you'll see users identified in the registry with things like "75726e3a7075626c696369643a49444e2b626f67757349442e6f72672b757365722b413157696e6e6572".  Eep!

Next concern, is that file results returned from algorithm execution will be brought into memory as Strings.  This is going to be a big concern in terms of memory usage, in the near future.  For now, fixing this would require some additional changes from Jiaan's side, and for our short-term purposes, we can go about things this way.  

Extending the RegistryClient to use streams, and hooking things up that way, shouldn't be a terribly big deal.  I have this marked in a couple places... including an "outstanding bugs/fixes needed" list that I'm growing.  So it will be addressed.

But for now, the new unit test passes, and while the agent isn't yet doing this automatically, the logic is there and it should be easy to add this in tomorrow.