Several hours spent in library hell.
Resolved the conflicting SLF4J logger issue that's been sticking around for a while.  Had to remove the lib for opensaml2-2.0.0.alpha1-wso2v1.jar, which I had previously assumed was a prereq for registryclientjar, but apparently is not.

sbt useful stuff:  clean-lib, clean-cache, clean-plugins.

Jiaan's new lib installed and I *think* it's working, testing more in the next iteration.

Eclipse .classpath file was updated, had previously hardcoded path.  This should likely be removed entirely, but with the relative paths it might make Aaron's build actually work.  Can potentially remove it if needed... it's a bit tricky to get it right, though.