Can now retrieve results from an algorithm's run via REST call.  Not heavily tested, but the basic cases work.
Agent is currently exporting handles to these results, but the portal is throwing them away.  Should be able to start on portal dealing with these handles tomorrow.


Example:  An algorithm ran and is now finished.  Client(portal) asks for the list of current algorithm runs and gets this back:

<algorithmRunList>
          <algorithmRun>
          <id>48f98acd-eb1e-4a7d-b9b9-7bbb3c204e99</id>
         <status>Finished</status><lastStatusChange> Mon Oct 03 16:46:31 EDT 2011 </lastStatusChange><workingDir>htrc-agent-working-dir/80287aca-171b-4438-8279-2971fa40776d</workingDir><results>
        <consoleStdout><href>/agent/urn:publicid:IDN+cilogon.org+user+A595/algorithm/48f98acd-eb1e-4a7d-b9b9-7bbb3c204e99/result/console/stdout</href></consoleStdout><file><href>/agent/urn:publicid:IDN+cilogon.org+user+A595/algorithm/48f98acd-eb1e-4a7d-b9b9-7bbb3c204e99/result/file/out-932197d3-1019-49a4-993b-8a50b133d288.txt</href></file>
       </results>
         </algorithmRun>
    </algorithmRunList>


Client can now call the following agent REST endpoints to get the output from the STDOUT console as well as the listed file output.  Examples of doing this from the command line:

 wget  http://coffeetree.cs.indiana.edu:41567/agent/urn:publicid:IDN+cilogon.org+user+A595/algorithm/48f98acd-eb1e-4a7d-b9b9-7bbb3c204e99/result/file/out-932197d3-1019-49a4-993b-8a50b133d288.txt -O -
 
 wget  http://coffeetree.cs.indiana.edu:41567/agent/urn:publicid:IDN+cilogon.org+user+A595/algorithm/48f98acd-eb1e-4a7d-b9b9-7bbb3c204e99/result/console/stdout -O -  
 
Although I'm not capturing STDERR right now, there's wiring that will export STDERR output when we start doing that.  It would look similar to the above, except the last part of the endpoint would be "console/stderr."