WARNING WARNING WARNING: MAJOR CHANGES NOT TESTED WITH PORTAL
Created a new class: RegistryActor. Goal is to move all registry functionality into this actor so that registry errors can be handled in one place. 

While the previous registry helper is still present agent and algorithm code are using the new actor.

The one major exception is that there are agent functions to get cassandra / solr / registry EPR's. These call the registry client directly. A "ToDo" is to decide if these are actually necessary and if so, move them to the registry actor.

A "Loggable" trait has also been added. Extending it provides access to a logger.