Looks like everything now works correctly.
A user can go to the portal, view collections and algorithms, run an algorithm, view status information and then view the result.

I concluded (possibly incorrectly) that the postResults methods are actually working. The portal error I view is a lack of xml formatting / style information. The results text file is viewable without issue from the web.

An important note on running the new agent is that since we have moved to akka 1.2 there needs to be an sbt "clean" and then "update" before "jetty-run". An inconvienence is that akka 1.1.2 libraries are in the subversion repository so until those are cleaned out this clean and update needs to be run after every svn up (assuming the svn up hits the libs folders).

In the process of getting this all working registry error handling was implemented. The current algorithm is as follows:

1) An information request is received by the registry actor.
2) The registry actor checks to see if the registry is alive by fetching the SolrURI
3) If the registry is dead it is restarted.
4) A spawn call is made to create a child actor tasked with responding to the query (frees RegistryActor to handle next query)
5) The child makes the registry access call.
6) If it fails, child retries up to 10 times.
7) Child responds to information request with the result of the registry access.

This is a somewhat naive approach. The issue is that the registry actor delegates registry interactions to children to avoid blocking on registry IO. Making this status check call pre spawn makes the registry actor block on IO. The proper solution, to be implemented at a future date, is to not have this keep-alive checking the registry before spawn.

Instead when a message is handled in a spawn block if there is a registry problem the message being handled is "requeued" in the registry actor's mailbox. If the registry actor notices a requeue it concludes the registry needs a restart does the restart.

This algorithm would be:

1) An information request is received by the registry actor.
  a) If a requeue restart registry
2) The registry actor spawns a child to process the message.
3) The child makes the registry access call.
4) If it fails, retry a few times.
5) If retires fail requeue the message being processed in the registry actor's mailbox.
6) Else send result.