Added a new RequestValidityChecker interface for checking the validity of each request. Added two implementations of RequestValidtyChecker, one for validating requests for volumes, the other for validating requests for pages.
These validity checkers now centralize various checks on the requests, including check for policy violations, as well as existence of requested IDs.  Earlier the code took a lazy approach that it would only know if an ID exists in cassandra or not by the time reading it out, this resulted in some zip files being corrupt as a non-existing ID is encountered later in the list of requested IDs, and by the time the exception is throw, the zip stream has already been committed.  the exception disrupts the output to the zip stream.  the validity checkers now take an active approach that they first check to make sure all IDs exist and throw exceptions if any one is missing.  This shall virtually eliminate the chance of generating corrupt zip files due to bad requests.